Discovering ok + IOSampling implementation

+++ org.openhab.binding.snavxbee2/src/main/java/org/openhab/binding/snavxbee2/devices/Tosr0xTparser.java
@@ -43,7 +43,13 @@ public class Tosr0xTparser {
 
             ChannelActionToPerform channelActionToPerform = new ChannelActionToPerform();
             channelActionToPerform.setChannelUIDToUpdate(new ChannelUID(thingUIDToUpdate + ":temperature"));
-            channelActionToPerform.setState(new DecimalType(Float.valueOf(new String(xbeeMessage.getData()))));
+            channelActionToPerform.setState(new DecimalType(Double.valueOf(new String(xbeeMessage.getData()))));
+
+            // DecimalFormat twoDForm = new DecimalFormat("#.##");
+            // logger.debug(" Temp ::::: {} ", Double.valueOf(twoDForm.format(new String(xbeeMessage.getData()))));
+            // channelActionToPerform
+            // .setState(new DecimalType(Double.valueOf(twoDForm.format(new String(xbeeMessage.getData())))));
+
             this.listOfChannelActionToPerform.add(channelActionToPerform);
 
         }
